Split debugging material into separate topic
This required a second level of subheadings, which indicates it
should be its own topic. One thing that's missing from this topic
right now is a set of challenges. I added a single challenge, which
I have done while teaching in the past, but additional challenges
would be good too.